+++ filesystem.go
@@ -2,6 +2,7 @@ package filesystem
 
 import (
 	"fmt"
+	"io"
 
 	"github.com/SevenOfSpades/go-just-options"
 )
@@ -20,8 +21,12 @@ func New(opts ...options.Option) (Filesystem, error) {
 	if err != nil {
 		return nil, fmt.Errorf("filesystem initialization failed: %w", err)
 	}
+	optCheckIfExistsHandler, err := options.ReadOrDefault[CheckIfExistsHandlerFunc](opt, optionCheckIfExistsHandler, checkIfExistsDefaultHandler)
+	if err != nil {
+		return nil, fmt.Errorf("filesystem initialization failed: %w", err)
+	}
 
-	return newFilesystem(optReadContentOfHandler, optStreamContentOfHandler)
+	return newFilesystem(optReadContentOfHandler, optStreamContentOfHandler, optCheckIfExistsHandler)
 }
 
 // ReadContentOf will return entire content of file from provided path.
@@ -34,15 +39,24 @@ func ReadContentOf(fs Filesystem, path string) (Content, error) {
 	return res, nil
 }
 
-// StreamContentOf will return pointer to struct implementing Streamer interface.
+// StreamContentOf will return pointer to struct implementing io.ReadCloser interface.
 // Implementation for this specific handler may or may not stream content directly from it source.
 // Some packages or sources for files require preloading entire content before performing any operation on it.
 //
 // If file does not exist it will return ErrFileNotFound error.
-func StreamContentOf(fs Filesystem, path string) (Streamer, error) {
+func StreamContentOf(fs Filesystem, path string) (io.ReadCloser, error) {
 	res, err := fs.handleStreamContentOf(path)
 	if err != nil {
 		return nil, fmt.Errorf("failed to attach reader to %s: %w", path, err)
+	}
+	return res, nil
+}
+
+// CheckIfExists will verify if file/directory exists on provided path.
+func CheckIfExists(fs Filesystem, path string) (bool, error) {
+	res, err := fs.handleCheckIfExists(path)
+	if err != nil {
+		return false, fmt.Errorf("failed to verify existence of %s: %w", path, err)
 	}
 	return res, nil
 }