+++ handlers.go
@@ -10,7 +10,10 @@ type (
 	ReadContentOfHandlerFunc func(string) (Content, error)
 
 	// StreamContentOfHandlerFunc is expected to be provided for as handler for StreamContentOf.
-	StreamContentOfHandlerFunc func(string) (Streamer, error)
+	StreamContentOfHandlerFunc func(string) (io.ReadCloser, error)
+
+	// CheckIfExistsHandlerFunc is expected to be provided for as handler for CheckIfExistsHandlerFunc.
+	CheckIfExistsHandlerFunc func(string) (bool, error)
 )
 
 func readContentOfDefaultHandler(path string) (Content, error) {
@@ -25,7 +28,7 @@ func readContentOfDefaultHandler(path string) (Content, error) {
 	return res, nil
 }
 
-func streamContentOfDefaultHandler(path string) (Streamer, error) {
+func streamContentOfDefaultHandler(path string) (io.ReadCloser, error) {
 	f, err := os.OpenFile(path, os.O_RDONLY, os.FileMode(0600)) //nolint:gosec
 	if err != nil {
 		if os.IsNotExist(err) {
@@ -34,4 +37,14 @@ func streamContentOfDefaultHandler(path string) (Streamer, error) {
 		return nil, err
 	}
 	return f, nil
+}
+
+func checkIfExistsDefaultHandler(path string) (bool, error) {
+	if _, err := os.Stat(path); err != nil {
+		if os.IsNotExist(err) {
+			return false, nil
+		}
+		return false, err
+	}
+	return true, nil
 }