+++ types.go
@@ -7,13 +7,8 @@ import (
 
 var ErrFileNotFound = errors.New("file not found")
 
-type (
-	Streamer interface {
-		io.Reader
-		io.Closer
-	}
-	Filesystem interface {
-		handleReadContentOf(string) (Content, error)
-		handleStreamContentOf(string) (Streamer, error)
-	}
-)
+type Filesystem interface {
+	handleReadContentOf(string) (Content, error)
+	handleStreamContentOf(string) (io.ReadCloser, error)
+	handleCheckIfExists(string) (bool, error)
+}