+++ CHANGELOG.md
@@ -5,8 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.2] - 2023-10-12
+
+* **Introduce `filesystem.CheckIfExists` function.**
+* Add `YAMLDecode` function to `filesystem.Content` type.
+* Replace `Streamer` interface with `io.ReadCloser`.
+
 ## [0.0.1] - 2023-10-10
 
- * Release to GitHub
+* Release to GitHub
 
+[0.0.2]: https://github.com/SevenOfSpades/go-wrapped-filesystem/releases/tag/v0.0.2
 [0.0.1]: https://github.com/SevenOfSpades/go-wrapped-filesystem/releases/tag/v0.0.1